# backend/openapi.yaml
openapi: 3.1.0
info:
  title: transx API
  description: transx API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /refresh:
    post:
      tags:
        - auth
      summary: Rotate refresh token and return a new access + refresh pair
      description: Rotate refresh token and return a new access + refresh pair
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DtoRefreshCommand'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DtoLoginResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
  /session:
    post:
      tags:
        - auth
      summary: Validate refresh token without rotating it (RR BFF auth-gate)
      description: Validate refresh token without rotating it (RR BFF auth-gate)
      operationId: session
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DtoRefreshCommand'
        required: true
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandlersErrorResponse'
...
components:
  schemas:
...
    DtoLoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
        userId:
          type: string
        userName:
          type: string
...
    DtoRefreshCommand:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
...
    HandlersErrorResponse:
      type: object
      properties:
        error:
          type: string
        requestId:
          type: string
        traceId:
          type: string